+++ src/components/queryBuilder/views/LogsQueryBuilder.tsx
@@ -32,6 +32,7 @@ interface LogsQueryBuilderState {
   timeColumn?: SelectedColumn;
   logLevelColumn?: SelectedColumn;
   messageColumn?: SelectedColumn;
+  labelsColumn?: SelectedColumn;
   // liveView: boolean;
   orderBy: OrderBy[];
   limit: number;
@@ -50,11 +51,13 @@ export const LogsQueryBuilder = (props: LogsQueryBuilderProps) => {
     timeColumn: getColumnByHint(builderOptions, ColumnHint.Time),
     logLevelColumn: getColumnByHint(builderOptions, ColumnHint.LogLevel),
     messageColumn: getColumnByHint(builderOptions, ColumnHint.LogMessage),
+    labelsColumn: getColumnByHint(builderOptions, ColumnHint.LogLabels),
     selectedColumns: builderOptions.columns?.filter(c => (
       // Only select columns that don't have their own box
       c.hint !== ColumnHint.Time &&
       c.hint !== ColumnHint.LogLevel &&
-      c.hint !== ColumnHint.LogMessage
+      c.hint !== ColumnHint.LogMessage &&
+      c.hint !== ColumnHint.LogLabels
     )) || [],
     // liveView: builderOptions.meta?.liveView || false,
     filters: builderOptions.filters || [],
@@ -74,6 +77,9 @@ export const LogsQueryBuilder = (props: LogsQueryBuilderProps) => {
     }
     if (next.messageColumn) {
       nextColumns.push(next.messageColumn);
+    }
+    if (next.labelsColumn) {
+      nextColumns.push(next.labelsColumn);
     }
 
     builderOptionsDispatch(setOptions({
@@ -154,6 +160,17 @@ export const LogsQueryBuilder = (props: LogsQueryBuilderProps) => {
           columnHint={ColumnHint.LogMessage}
           label={labels.logMessageColumn.label}
           tooltip={labels.logMessageColumn.tooltip}
+        />
+        <ColumnSelect
+          disabled={builderState.otelEnabled}
+          allColumns={allColumns}
+          selectedColumn={builderState.labelsColumn}
+          invalid={!builderState.labelsColumn}
+          onColumnChange={onOptionChange('labelsColumn')}
+          columnHint={ColumnHint.LogLabels}
+          label={labels.logLabelsColumn.label}
+          tooltip={labels.logLabelsColumn.tooltip}
+          inline
         />
         {/* <Switch
           value={builderState.liveView}