+++ src/labels.ts
@@ -306,6 +306,10 @@ export default {
         label: 'Message',
         tooltip: 'Column that contains the log message'
       },
+      logLabelsColumn: {
+        label: 'Labels',
+        tooltip: 'A column with a key/value structure for log labels'
+      },
       liveView: {
         label: 'Live View',
         tooltip: 'Enable to update logs in real time'
@@ -408,6 +412,7 @@ export default {
 
       [ColumnHint.LogLevel]: 'Level',
       [ColumnHint.LogMessage]: 'Message',
+      [ColumnHint.LogLabels]: 'Labels',
 
       [ColumnHint.TraceId]: 'Trace ID',
       [ColumnHint.TraceSpanId]: 'Span ID',